Move validation processor to its own module
